Remove old version of auth
