+++ src/auth/app.py
@@ -1,8 +1,7 @@
 import uvicorn
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from src.auth.routers.auth import auth_router as arv1
-from src.auth.routers.new_auth import auth_router as arv2
+from src.auth.routers.auth import auth_router
 
 
 app = FastAPI()
@@ -32,8 +31,7 @@ app.add_middleware(
 )
 
 
-app.include_router(arv1)
-app.include_router(arv2)
+app.include_router(auth_router)
 
 
 if __name__ == "__main__":
